+++ mycelium-cli/src/lib.rs
@@ -4,6 +4,7 @@ mod message;
 mod peer;
 mod proxy;
 mod routes;
+mod stats;
 
 pub use inspect::inspect;
 #[cfg(feature = "message")]
@@ -15,3 +16,4 @@ pub use proxy::{
 pub use routes::{
     list_fallback_routes, list_no_route_entries, list_queried_subnets, list_selected_routes,
 };
+pub use stats::list_packet_stats;