+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.2] - 2026-02-09
+
 ### Added
 
 - You can now limit peer discovery to specific interfaces by specifying the
@@ -686,4 +688,5 @@ This is a breaking change, check the main README file for update info.
 [0.6.2]: https://github.com/threefoldtech/mycelium/compare/v0.6.1...v0.6.2
 [0.7.0]: https://github.com/threefoldtech/mycelium/compare/v0.6.2...v0.7.0
 [0.7.1]: https://github.com/threefoldtech/mycelium/compare/v0.7.0...v0.7.1
-[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.7.1...HEAD
+[0.7.2]: https://github.com/threefoldtech/mycelium/compare/v0.7.1...v0.7.2
+[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.7.2...HEAD